# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x0tta6bl4 Mesh Network API",
    "version": "3.2.1",
    "description": "\n# x0tta6bl4 Mesh Network API\n\nSelf-healing mesh network node with MAPE-K autonomic loop and Kimi K2.5 Agent Swarm.\n\n## Features\n\n- **Post-Quantum Cryptography**: ML-KEM-768, ML-DSA-65 (NIST FIPS 203/204)\n- **SPIFFE/SPIRE Integration**: Zero Trust workload identity with mTLS\n- **Agent Swarm**: Parallel task execution with up to 100 agents\n- **PARL**: Parallel Agent Reinforcement Learning for 4.5x speedup\n- **Vision Module**: Visual analysis for mesh topology and anomaly detection\n\n## Authentication\n\nMost endpoints require `X-Admin-Token` header for authentication.\n\n## Rate Limiting\n\n- Standard endpoints: 100 requests/minute\n- Task submission: 1000 requests/minute\n- Swarm creation: 10 requests/minute\n- Scaling: 5 requests/minute\n\n## Error Handling\n\nAll errors follow RFC 7807 Problem Details format:\n```json\n{\n  \"type\": \"https://x0tta6bl4.mesh/errors/swarm-not-found\",\n  \"title\": \"Swarm Not Found\",\n  \"status\": 404,\n  \"detail\": \"Swarm with ID 'abc123' does not exist\"\n}\n```\n        ",
    "contact": {
      "name": "x0tta6bl4 Support",
      "email": "support@x0tta6bl4.mesh",
      "url": "https://x0tta6bl4.mesh"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-logo": {
      "url": "/logo.svg",
      "altText": "x0tta6bl4 Logo"
    }
  },
  "paths": {
...
    "/api/v3/swarm/create": {
      "post": {
        "tags": [
          "swarm"
        ],
        "summary": "Create Swarm",
        "description": "Create a new agent swarm.\n\nCreates a swarm with specified number of agents and capabilities.\nUses PARL for parallel task execution with up to 4.5x speedup.",
        "operationId": "create_swarm_api_v3_swarm_create_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwarmCreateRequest"
              },
              "examples": {
                "basic": {
                  "$ref": "#/components/examples/SwarmCreateRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwarmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/status": {
      "get": {
        "tags": [
          "swarm"
        ],
        "summary": "Get Swarm Status",
        "description": "Get current status of a swarm.",
        "operationId": "get_swarm_status_api_v3_swarm__swarm_id__status_get",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/tasks": {
      "post": {
        "tags": [
          "swarm"
        ],
        "summary": "Submit Task",
        "description": "Submit a task for execution by the swarm.",
        "operationId": "submit_task_api_v3_swarm__swarm_id__tasks_post",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmitRequest"
              },
              "examples": {
                "mesh_analysis": {
                  "$ref": "#/components/examples/TaskSubmitRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/tasks/batch": {
      "post": {
        "tags": [
          "swarm"
        ],
        "summary": "Submit Tasks Batch",
        "description": "Submit multiple tasks for parallel execution.",
        "operationId": "submit_tasks_batch_api_v3_swarm__swarm_id__tasks_batch_post",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TaskSubmitRequest"
                },
                "title": "Tasks"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/tasks/{task_id}": {
      "get": {
        "tags": [
          "swarm"
        ],
        "summary": "Get Task Status",
        "description": "Get status of a specific task.",
        "operationId": "get_task_status_api_v3_swarm__swarm_id__tasks__task_id__get",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/agents": {
      "get": {
        "tags": [
          "swarm"
        ],
        "summary": "List Agents",
        "description": "List all agents in a swarm.",
        "operationId": "list_agents_api_v3_swarm__swarm_id__agents_get",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/agents/{agent_id}": {
      "get": {
        "tags": [
          "swarm"
        ],
        "summary": "Get Agent Status",
        "description": "Get detailed status of a specific agent.",
        "operationId": "get_agent_status_api_v3_swarm__swarm_id__agents__agent_id__get",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/agents/{agent_id}/control": {
      "post": {
        "tags": [
          "swarm"
        ],
        "summary": "Control Agent",
        "description": "Control agent state (pause, resume, restart, terminate).",
        "operationId": "control_agent_api_v3_swarm__swarm_id__agents__agent_id__control_post",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/swarm/{swarm_id}/scale": {
      "post": {
        "tags": [
          "swarm"
        ],
        "summary": "Scale Swarm",
        "description": "Scale the swarm up or down.",
        "operationId": "scale_swarm_api_v3_swarm__swarm_id__scale_post",
        "parameters": [
          {
            "name": "swarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swarm Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScaleRequest"
              },
              "examples": {
                "scale_up": {
                  "$ref": "#/components/examples/ScaleUpRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/swarm": {
      "get": {
        "tags": [
          "swarm"
        ],
        "summary": "List Swarms",
        "description": "List all active swarms.",
        "operationId": "list_swarms_api_v3_swarm_get",
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token",
        "description": "Admin authentication token"
      },
      "mTLS": {
        "type": "mutualTLS",
        "description": "Mutual TLS authentication with SPIFFE SVID"
      }
    },
    "examples": {
...
      "ScaleUpRequest": {
        "summary": "Scale up swarm",
        "value": {
          "action": "scale_up",
          "num_agents": 5,
          "reason": "Increased load"
        }
      }
    }
  },
  "servers": [
...
}